+++ web/src/modules/travel-authorizations/components/MyTravelAuthorizationsTable.vue
@@ -132,7 +132,7 @@ export default {
     goToFormDetails(form) {
       const formId = form.id
       if (form.status === "draft") {
-        this.$router.push({ name: "TravelFormEdit-DetailsTab", params: { formId } })
+        this.$router.push({ name: "TravelAuthorizationEdit-DetailsTab", params: { formId } })
       } else {
         this.$router.push({
           name: "TravelAuthorizationRead-DetailsTab",
