+++ web/src/modules/travel-authorizations/pages/travel-authorization-edit/details-tab/DetailsFormCard.vue
