+++ web/src/modules/travel-authorizations/pages/travel-authorization-edit/details-tab/approvals-form-card/SubmitToSupervisorButton.vue
