+++ web/src/modules/travel-authorizations/pages/travel-authorization-edit/details-tab/details-form-card/MultiDestinationStopsSection.vue
