+++ web/src/modules/travel-authorizations/pages/travel-authorization-edit/estimate-tab/EstimateCreateDialog.vue
@@ -3,7 +3,7 @@
     v-model="showDialog"
     max-width="500px"
   >
-    <template v-slot:activator="{ on, attrs }">
+    <template #activator="{ on, attrs }">
       <v-btn
         color="primary"
         dark
