+++ web/src/modules/travel-authorizations/pages/travel-form-list/CreateTravelAuthorizationButton.vue
@@ -20,7 +20,7 @@ export default {
     ...mapActions("travelAuthorizations", ["create"]),
     goToFormDetails(form) {
       const formId = form.id
-      this.$router.push({ name: "TravelFormEdit-DetailsTab", params: { formId } })
+      this.$router.push({ name: "TravelAuthorizationEdit-DetailsTab", params: { formId } })
     },
     createAndGoToFormDetails() {
       this.loadingCreatingForm = true
